api/python-ai-agents/requirements.txt
fastapi
uvicorn
phidata
google-generativeai
python-multipart
duckduckgo-search
python-dotenv
pydantic
pillow